add graph_gen and performance tests
