--- src/__init__.py
@@ -1,1 +1,0 @@
-from .loader import load
