+++ src/graph_gen/__init__.py
@@ -1,0 +1,2 @@
+from .free_scale_network import generate_free_scale_graph
+from .graph2txt import graph2txt
