+++ src/graph_gen/graphs/directed_free_scale_net_500_1.txt
@@ -1,0 +1,500 @@
+0 a 0
+1 b 0
+2 b 0
+3 c 1
+4 a 1
+5 b 0
+6 b 0
+7 d 0
+8 c 0
+9 a 6
+10 b 6
+11 d 0
+12 c 8
+13 d 2
+14 b 9
+15 a 0
+16 b 11
+17 b 4
+18 c 6
+19 d 9
+20 c 11
+21 b 11
+22 c 9
+23 a 0
+24 c 11
+25 c 0
+26 c 1
+27 c 20
+28 a 23
+29 c 0
+30 d 0
+31 b 10
+32 b 20
+33 d 0
+34 c 9
+35 b 20
+36 b 0
+37 d 27
+38 a 28
+39 b 0
+40 a 12
+41 a 4
+42 b 21
+43 c 9
+44 d 36
+45 d 0
+46 c 33
+47 c 23
+48 a 41
+49 c 0
+50 c 5
+51 c 8
+52 a 33
+53 b 6
+54 b 20
+55 d 0
+56 b 9
+57 a 0
+58 b 4
+59 c 11
+60 b 0
+61 c 0
+62 d 29
+63 a 9
+64 d 3
+65 b 35
+66 b 10
+67 b 43
+68 d 8
+69 b 1
+70 c 52
+71 a 0
+72 b 11
+73 c 0
+74 b 8
+75 b 8
+76 b 1
+77 a 33
+78 a 33
+79 b 29
+80 c 6
+81 a 1
+82 b 30
+83 a 56
+84 a 6
+85 a 6
+86 c 23
+87 c 11
+88 c 9
+89 b 73
+90 d 0
+91 d 1
+92 c 71
+93 c 1
+94 a 0
+95 a 63
+96 c 68
+97 d 93
+98 d 11
+99 b 73
+100 a 16
+101 c 63
+102 b 13
+103 a 57
+104 a 61
+105 b 20
+106 c 53
+107 b 4
+108 a 17
+109 b 20
+110 c 68
+111 c 58
+112 b 19
+113 b 33
+114 b 6
+115 b 0
+116 a 87
+117 a 0
+118 b 0
+119 d 78
+120 c 35
+121 b 29
+122 c 96
+123 b 44
+124 d 82
+125 d 8
+126 a 35
+127 a 3
+128 b 12
+129 d 17
+130 c 128
+131 b 29
+132 a 51
+133 d 130
+134 c 4
+135 a 87
+136 d 34
+137 d 65
+138 a 0
+139 b 103
+140 a 1
+141 a 138
+142 a 8
+143 a 29
+144 b 138
+145 a 29
+146 d 0
+147 a 129
+148 d 59
+149 a 12
+150 a 1
+151 a 42
+152 c 146
+153 b 76
+154 c 117
+155 b 154
+156 c 11
+157 d 1
+158 b 0
+159 c 63
+160 b 0
+161 a 90
+162 d 3
+163 a 87
+164 b 128
+165 b 2
+166 c 89
+167 d 51
+168 d 39
+169 c 132
+170 d 56
+171 b 0
+172 a 17
+173 a 20
+174 b 27
+175 a 0
+176 a 35
+177 d 8
+178 c 156
+179 a 20
+180 b 37
+181 b 117
+182 a 27
+183 c 25
+184 b 69
+185 d 66
+186 b 4
+187 b 76
+188 a 175
+189 b 81
+190 c 146
+191 b 103
+192 c 130
+193 c 0
+194 b 10
+195 b 6
+196 a 32
+197 a 29
+198 b 114
+199 a 4
+200 d 4
+201 a 182
+202 c 103
+203 c 49
+204 c 145
+205 c 1
+206 c 77
+207 a 173
+208 b 56
+209 b 40
+210 d 139
+211 a 3
+212 d 145
+213 d 2
+214 c 3
+215 c 73
+216 d 49
+217 a 204
+218 c 128
+219 b 70
+220 c 1
+221 c 63
+222 a 11
+223 c 19
+224 c 186
+225 c 128
+226 c 42
+227 c 63
+228 c 10
+229 d 12
+230 b 20
+231 a 115
+232 d 31
+233 b 24
+234 d 46
+235 b 0
+236 a 132
+237 a 35
+238 c 30
+239 b 6
+240 a 10
+241 a 0
+242 b 132
+243 d 3
+244 d 136
+245 c 163
+246 c 132
+247 c 6
+248 a 195
+249 b 0
+250 a 128
+251 c 30
+252 b 38
+253 c 6
+254 c 170
+255 b 1
+256 b 49
+257 d 137
+258 a 41
+259 a 76
+260 d 29
+261 a 0
+262 d 36
+263 b 93
+264 a 110
+265 b 143
+266 c 257
+267 a 137
+268 d 151
+269 c 0
+270 c 185
+271 b 1
+272 a 46
+273 c 0
+274 a 138
+275 b 128
+276 b 44
+277 b 87
+278 a 46
+279 b 1
+280 c 107
+281 d 12
+282 d 11
+283 b 281
+284 a 93
+285 b 71
+286 c 42
+287 c 29
+288 b 27
+289 c 177
+290 b 65
+291 c 0
+292 b 17
+293 d 22
+294 a 154
+295 a 130
+296 d 0
+297 b 29
+298 b 61
+299 b 187
+300 c 36
+301 b 29
+302 b 271
+303 b 0
+304 a 283
+305 a 20
+306 b 271
+307 d 4
+308 c 20
+309 d 0
+310 b 134
+311 d 132
+312 a 73
+313 c 12
+314 b 157
+315 c 6
+316 d 1
+317 a 160
+318 b 270
+319 b 203
+320 a 128
+321 c 41
+322 c 301
+323 d 128
+324 b 0
+325 d 36
+326 a 172
+327 a 297
+328 d 1
+329 c 16
+330 b 202
+331 c 5
+332 c 163
+333 b 126
+334 a 6
+335 a 70
+336 c 115
+337 d 183
+338 a 0
+339 d 49
+340 d 96
+341 b 20
+342 d 93
+343 a 335
+344 c 228
+345 b 11
+346 b 258
+347 a 146
+348 a 183
+349 a 249
+350 c 148
+351 a 56
+352 c 156
+353 d 56
+354 b 86
+355 c 80
+356 b 87
+357 c 16
+358 b 117
+359 c 107
+360 a 156
+361 d 63
+362 d 144
+363 b 0
+364 d 87
+365 d 11
+366 a 10
+367 d 364
+368 d 249
+369 b 65
+370 b 11
+371 b 232
+372 d 128
+373 c 250
+374 c 8
+375 c 8
+376 c 70
+377 a 180
+378 a 37
+379 b 151
+380 a 1
+381 d 201
+382 b 21
+383 b 0
+384 b 4
+385 a 334
+386 d 26
+387 a 369
+388 b 184
+389 d 87
+390 d 66
+391 b 33
+392 a 145
+393 c 222
+394 a 11
+395 a 135
+396 b 93
+397 c 0
+398 c 1
+399 c 3
+400 c 280
+401 b 174
+402 c 186
+403 d 305
+404 a 383
+405 d 154
+406 b 288
+407 c 251
+408 c 165
+409 d 16
+410 b 304
+411 d 11
+412 b 19
+413 d 369
+414 b 61
+415 a 95
+416 a 70
+417 c 99
+418 c 195
+419 d 296
+420 b 419
+421 c 0
+422 d 338
+423 d 233
+424 c 36
+425 c 243
+426 d 106
+427 b 109
+428 b 154
+429 a 194
+430 c 35
+431 d 93
+432 c 29
+433 c 218
+434 d 40
+435 a 152
+436 b 251
+437 a 87
+438 c 11
+439 d 251
+440 c 289
+441 b 4
+442 b 138
+443 c 9
+444 d 317
+445 b 46
+446 b 257
+447 b 381
+448 a 0
+449 b 87
+450 d 411
+451 a 179
+452 c 10
+453 b 262
+454 b 175
+455 b 187
+456 b 50
+457 c 51
+458 b 138
+459 b 103
+460 b 387
+461 d 360
+462 a 260
+463 b 29
+464 c 155
+465 c 104
+466 d 138
+467 c 381
+468 b 65
+469 a 56
+470 c 431
+471 b 408
+472 a 202
+473 b 67
+474 a 335
+475 c 290
+476 a 328
+477 d 128
+478 a 405
+479 a 138
+480 a 465
+481 c 104
+482 a 29
+483 c 238
+484 b 13
+485 c 465
+486 d 103
+487 a 0
+488 d 73
+489 c 329
+490 b 128
+491 d 0
+492 b 36
+493 d 52
+494 b 208
+495 c 55
+496 b 10
+497 d 417
+498 b 46
+499 b 35
