+++ src/performance/__init__.py
@@ -1,0 +1,1 @@
+from .graph_testing import test_performance_on_graph, test_performance_on_suite
