+++ src/performance/main.py
@@ -1,12 +1,6 @@
-import hashlib
-from pprint import pprint
-
+import datetime
 import pandas as pd
-
-import redis
-
-COMMON_CMD = 'graph.cfg'
-ALGO_LIST = ['cpu1', 'cpu3']
+from src.performance import test_performance_on_suite
 
 RDF_GRAMMAR_PATH = '/home/jblab/CFPQ-with-RedisGraph/CFPQ_Data/data/graphs/RDF/Grammars'
 WS_GRAMMAR_PATH = '/home/jblab/CFPQ-with-RedisGraph/CFPQ_Data/data/graphs/WorstCase/Grammars/Brackets.txt'
@@ -40,49 +34,9 @@ TEST_SUITE = [
     ('worstcase_512.txt', WS_GRAMMAR_PATH)
 ]
 
-r = redis.Redis()
 
+results = test_performance_on_suite(TEST_SUITE)
+results_df = pd.DataFrame(results)
 
-def test_performance_on_graph(graph_name, grammar_path):
-    print(f'Test {graph_name}')
-    res_query = [
-        [s.decode('UTF-8')
-         for s in r.execute_command(COMMON_CMD, algo, graph_name, grammar_path)]
-        for algo in ALGO_LIST
-    ]
-    res_dict = {
-        'iterations': [],
-        'control_sum': [],
-        'time': []
-    }
-
-    for algo, res in zip(ALGO_LIST, res_query):
-        res_dict['iterations'].append(res[1].split()[-1])
-        res_dict['control_sum'].append(", ".join(res[2:]))
-        res_dict['time'].append(res[0].split()[-1])
-
-    return res_dict
-
-
-def test_performance():
-    total_res = {
-        'name': [],
-        'grammar': []
-    }
-    for graph_name, grammar_path in TEST_SUITE:
-        res_test = test_performance_on_graph(graph_name, grammar_path)
-        for column in['iterations', 'control_sum', 'time']:
-            for value, algo in zip(res_test[column], ALGO_LIST):
-                key = f'{column}_{algo}'
-                total_res.setdefault(key, [])
-                total_res[key].append(value)
-        total_res['name'].append(graph_name)
-        total_res['grammar'].append(grammar_path.split('/')[-1])
-    return total_res
-
-
-xs = test_performance()
-pprint(xs, indent=4)
-ys = pd.DataFrame(xs)
-ys.to_csv('test_results.csv')
-print(ys)
+now = datetime.datetime.now().strftime('"%Y-%m-%d_%H:%M"')
+results_df.to_csv(f'results/results_{now}.csv')
