+++ src/redis_loader/__init__.py
@@ -1,0 +1,1 @@
+from .loader import load
